refactor(ndk): split ndk.ts into ndk/ modules (cache, connection, subscribe, index)
